[Kernel VFS ProcFS]: refactoring the procfs module
